feat: #1325 - PaperMigrator first pass complete - further changes after PR feedback
