+++ services/net/contentmigration/migrators/ContentMigrator.cs
@@ -14,61 +14,6 @@ using TNO.Services.ContentMigration.Sources.Oracle;
 namespace TNO.Services.ContentMigration.Migrators;
 
 /// <summary>
-/// Interface for ContentMigrator implementations
-/// </summary>
-public interface IContentMigrator
-{
-    /// <summary>
-    /// which Ingests this Migrator supports
-    /// </summary>
-    IEnumerable<string> SupportedIngests { get; }
-
-    /// <summary>
-    ///
-    /// </summary>
-    /// <returns></returns>
-    Expression<Func<NewsItem, bool>> GetBaseFilter();
-
-    /// <summary>
-    ///
-    /// </summary>
-    /// <param name="lookup"></param>
-    /// <param name="newsItemSource"></param>
-    /// <returns></returns>
-    SourceModel? GetSourceMapping(IEnumerable<SourceModel> lookup, string newsItemSource);
-
-    /// <summary>
-    ///
-    /// </summary>
-    /// <param name="lookup"></param>
-    /// <param name="newsItem"></param>
-    /// <returns></returns>
-    ProductModel? GetProductMapping(IEnumerable<ProductModel> lookup, NewsItem newsItem);
-
-    /// <summary>
-    /// Creates an Clip ContentReferenceModel from a NewsItem
-    /// </summary>
-    /// <param name="source"></param>
-    /// <param name="topic"></param>
-    /// <param name="newsItem"></param>
-    /// <param name="uid"></param>
-    /// <returns></returns>
-    ContentReferenceModel CreateContentReference(SourceModel source, string topic, NewsItem newsItem, string uid);
-
-    /// <summary>
-    /// Creates a SourceContent item
-    /// </summary>
-    /// <param name="lookups"></param>
-    /// <param name="source"></param>
-    /// <param name="product"></param>
-    /// <param name="contentType"></param>
-    /// <param name="newsItem"></param>
-    /// <param name="referenceUid"></param>
-    /// <returns></returns>
-    SourceContent? CreateSourceContent(LookupModel lookups, SourceModel source, ProductModel product, ContentType contentType, NewsItem newsItem, string referenceUid);
-}
-
-/// <summary>
 ///
 /// </summary>
 /// <typeparam name="TOptions"></typeparam>
@@ -120,7 +65,8 @@ public abstract class ContentMigrator<TOptions> : IContentMigrator
         this.Api = api;
         this.Options = options.Value;
         this.Logger = logger;
-        this.MigratorOptions = migratorOptions.Get(this.GetType().Name); // "ClipMigrator"
+        // as a convention the Named Option should be the same as the class name which needs to consume it
+        this.MigratorOptions = migratorOptions.Get(this.GetType().Name);
     }
     #endregion
